requirements.txt
requests>=2.14
gitpython
